requirements.txt
backports-abc==0.5
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9
cssselect==1.1.0
decorator==4.4.2
enum34==1.1.10
Flask==1.1.2
future==0.18.2
futures==3.3.0
gevent==1.5.0
greenlet==0.4.15
hiredis==1.0.1
idna==2.9
ipaddress==1.0.23
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
psutil==5.7.0
pycparser==2.20
pyquery==1.4.1
PySocks==1.7.1
python-telegram==0.12.0
python-telegram-bot==12.6.1
redis==3.5.3
redisearch==0.9.0
requests==2.23.0
rmtest==0.7.0
singledispatch==3.4.0.3
six==1.14.0
tornado==5.1.1
urllib3==1.25.8
Werkzeug==1.0.1
